requirements.txt
python-dotenv
openai
chromadb
tiktoken
langchain
python-docx
pymupdf
requests
unstructured
matplotlib
pandas